modulacion dashboard por secciones en componentes
